changing data type for median

+++ data/src/main/java/com/proptiger/data/model/filter/JPAQueryBuilder.java
@@ -163,7 +163,7 @@ public class JPAQueryBuilder<T extends BaseModel> extends AbstractQueryBuilder<T
                 break;
             case MEDIAN:
                 Expression<Double> medianExpression = root.get(actualFieldName);
-                expression = criteriaBuilder.function("median", Integer.class, medianExpression);
+                expression = criteriaBuilder.function("median", Double.class, medianExpression);
                 break;
             case COUNT:
             	Expression<Number> countExpression;
